+++ introspect/functions.py
@@ -6187,64 +6187,6 @@ FUNCTIONS: Mapping[str, FunctionDecl] = dict([
          ),
          doc="Multiply transposed 3-by-3 matrix by vector: res = mat' * vec.",
      )),
-    ('mju_rotVecMat',
-     FunctionDecl(
-         name='mju_rotVecMat',
-         return_type=ValueType(name='void'),
-         parameters=(
-             FunctionParameterDecl(
-                 name='res',
-                 type=ArrayType(
-                     inner_type=ValueType(name='mjtNum'),
-                     extents=(3,),
-                 ),
-             ),
-             FunctionParameterDecl(
-                 name='vec',
-                 type=ArrayType(
-                     inner_type=ValueType(name='mjtNum', is_const=True),
-                     extents=(3,),
-                 ),
-             ),
-             FunctionParameterDecl(
-                 name='mat',
-                 type=ArrayType(
-                     inner_type=ValueType(name='mjtNum', is_const=True),
-                     extents=(9,),
-                 ),
-             ),
-         ),
-         doc='Deprecated, use mju_mulMatVec3(res, mat, vec).',
-     )),
-    ('mju_rotVecMatT',
-     FunctionDecl(
-         name='mju_rotVecMatT',
-         return_type=ValueType(name='void'),
-         parameters=(
-             FunctionParameterDecl(
-                 name='res',
-                 type=ArrayType(
-                     inner_type=ValueType(name='mjtNum'),
-                     extents=(3,),
-                 ),
-             ),
-             FunctionParameterDecl(
-                 name='vec',
-                 type=ArrayType(
-                     inner_type=ValueType(name='mjtNum', is_const=True),
-                     extents=(3,),
-                 ),
-             ),
-             FunctionParameterDecl(
-                 name='mat',
-                 type=ArrayType(
-                     inner_type=ValueType(name='mjtNum', is_const=True),
-                     extents=(9,),
-                 ),
-             ),
-         ),
-         doc='Deprecated, use mju_mulMatTVec3(res, mat, vec).',
-     )),
     ('mju_cross',
      FunctionDecl(
          name='mju_cross',